:recycle: Move all parts for Elimine in a dedicated folder
