+++ src/views/certificates/ppe2/components/elimine/Elimine.tsx
@@ -1,7 +1,7 @@
 import { Text } from '@react-pdf/renderer'
 import React from 'react'
 import { ProjectDataForCertificate } from '@modules/project/dtos'
-import { estSoumisAuxGFs } from '../helpers/estSoumisAuxGFs'
+import { estSoumisAuxGFs } from '../../helpers/estSoumisAuxGFs'
 import { MotifElimination } from './motifElimination'
 
 type ElimineProps = {
