+++ src/views/certificates/ppe2/components/elimine/index.ts
@@ -1,0 +1,1 @@
+export * from './Elimine'
