+++ src/views/certificates/ppe2/components/elimine/motifElimination/AuDessusDePcible.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ProjectDataForCertificate } from 'src/modules/project'
-import { formatNumber } from '../../helpers/formatNumber'
-import { getNoteThreshold } from '../../helpers/getNoteThreshold'
+import { formatNumber } from '../../../helpers/formatNumber'
+import { getNoteThreshold } from '../../../helpers/getNoteThreshold'
 
 type AuDessusDePcibleProps = {
   project: ProjectDataForCertificate
