+++ src/views/certificates/ppe2/components/elimine/motifElimination/MotifElimination.tsx
